fix: resolve ApplicationSet project field for RBAC checks
This fixes the issue where ApplicationSet RBAC checks were failing when
the project field contained templates like {{default "default" .project}}.

The problem was that the RBACName method was being called with raw
ApplicationSet objects that might have unprocessed template strings
in the project field. This caused RBAC enforcement to fail because
the project name couldn't be properly resolved.

The fix adds a new helper method getApplicationSetRBACName that:
1. Checks if the project field contains templates ({{...}})
2. If it does, uses the default project for RBAC checks
3. If the project field is empty, also uses the default project
4. Otherwise, uses the actual project name

This ensures that users with access to the default project can access
ApplicationSets with templated project fields, which is the expected
behavior.

Fixes: https://github.com/argoproj/argo-cd/issues/24561
Signed-off-by: Akhil Singh <singhakhil69@gmail.com>